Refactoring - moving pakages (WIP)
