+++ queries/src/main/scala/cz/cvut/fit/prl/scalaimplicit/queries/Main.scala
@@ -1,7 +1,5 @@
 package cz.cvut.fit.prl.scalaimplicit.queries
 
-import cz.cvut.fit.prl.scalaimplicit.matcher.PredefinedQueries
-
 object Main {
 
   def main(args: Array[String]): Unit = {
