+++ queries/src/main/scala/cz/cvut/fit/prl/scalaimplicit/queries/PredefinedQueries.scala
@@ -1,4 +1,4 @@
-package cz.cvut.fit.prl.scalaimplicit.matcher
+package cz.cvut.fit.prl.scalaimplicit.queries
 
 import java.nio.file.{Files, Paths}
 
@@ -7,10 +7,7 @@ import cz.cvut.fit.prl.scalaimplicit.core.extractor.representation.SlimRepresent
 import cz.cvut.fit.prl.scalaimplicit.core.extractor.serializers.HTMLSerializer
 import cz.cvut.fit.prl.scalaimplicit.core.extractor.serializers.HTMLSerializer.{TCFamily, TCItem}
 import cz.cvut.fit.prl.scalaimplicit.core.reports.{DefinitionSummary, ProjectMetadata, ProjectReport, SlimReport}
-import cz.cvut.fit.prl.scalaimplicit.queries.{OutputHelper, QueryEngine}
 import cz.cvut.fit.prl.scalaimplicit.queries.QueryEngine.CSFilterQuery
-
-import scala.collection.immutable
 
 object PredefinedQueries {
   def typeClassClassification() = {
